Split up setup into two plugins
