+++ lib/plugins/aws/setup/awsSetup.test.js
@@ -1,0 +1,1 @@
+// TODO: write tests
