+++ lib/plugins/setup/setup.test.js
@@ -16,11 +16,4 @@ describe('Setup', () => {
   describe('#constructor()', () => {
     it('should have commands', () => expect(setup.commands).to.be.not.empty);
   });
-
-  describe('#setup()', () => {
-    it('should throw error if user passed unsupported provider', () => {
-      setup.options = { provider: 'blah' };
-      expect(() => setup.setup()).to.throw(Error);
-    });
-  });
 });
